Move config stuff out to a module.
